+++ backup.sh
@@ -1,68 +1,64 @@
 #!/usr/bin/env bash
 #
-# description of the script
+# Backup script to save backup to repository and forgets old snapshots (execute manually or by CHECK_CRON schedule)
+# - restic backup
+# - restic forget
 
 set -eo pipefail
 
-#
+source /bin/log.sh
 
-lastLogfile="/var/log/backup-last.log"
-lastMailLogfile="/var/log/mail-last.log"
+lastLogfile="/var/log/restic-backup.log"
 backup_dir="/data"
 
-copyErrorLog() {
-  cp ${lastLogfile} /var/log/backup-error-last.log
-}
-
-logLast() {
-  echo "$1" >> ${lastLogfile}
-}
 
 healthcheck() {
     suffix=$1
     if [ -n "$HEALTHCHECK_URL" ]; then
-        echo "Reporting healthcheck $suffix ..."
+        log "INFO" "Reporting healthcheck $suffix ..."
         [[ ${1} == "/start" ]] && m="" || m=$(cat ${lastLogfile} | tail -n 100)
         curl -fSsL --retry 3 -X POST \
             --user-agent "seatable-restic/1.0.0" \
             --data-raw "$m" "${HEALTHCHECK_URL}${suffix}"
-        echo
         if [ $? != 0 ]; then
             echo "HEALTHCHECK_URL seems to be wrong..."
             exit 1
         fi
     else
-        echo "No HEALTHCHECK_URL provided. Skipping healthcheck."
+        log "DEBUG" "No HEALTHCHECK_URL provided. Skipping healthcheck."
     fi
 }
 
+# always execute /bin/pre-default.sh
+log "DEBUG" "Starting pre-default.sh"
 /bin/pre-default.sh
 
+# /hooks/pre-backup.sh
 if [ -f "/hooks/pre-backup.sh" ]; then
-    echo "Starting pre-backup script ..."
+    log "INFO" "Starting pre-backup script"
     /hooks/pre-backup.sh
 else
-    echo "Pre-backup script not found ..."
+    log "DEBUG" "Pre-backup script not found"
 fi
 
 start=`date +%s`
-rm -f ${lastLogfile} ${lastMailLogfile}
-echo "Starting Backup at $(date +"%Y-%m-%d %H:%M:%S")"
-echo "Starting Backup at $(date)" >> ${lastLogfile}
-logLast "BACKUP_CRON: ${BACKUP_CRON}"
-logLast "RESTIC_TAG: ${RESTIC_TAG}"
-logLast "RESTIC_FORGET_ARGS: ${RESTIC_FORGET_ARGS}"
-logLast "RESTIC_JOB_ARGS: ${RESTIC_JOB_ARGS}"
-logLast "RESTIC_REPOSITORY: ${RESTIC_REPOSITORY}"
-logLast ""
-logLast "Directory tree:"
-tree -a -P .exclude_from_backup -L 4 ${backup_dir} | tee -a "${lastLogfile}"
+log "INFO" "Starting Backup"
+log "DEBUG" "BACKUP_CRON: ${BACKUP_CRON}"
+log "DEBUG" "RESTIC_TAG: ${RESTIC_TAG}"
+log "DEBUG" "RESTIC_FORGET_ARGS: ${RESTIC_FORGET_ARGS}"
+log "DEBUG" "RESTIC_JOB_ARGS: ${RESTIC_JOB_ARGS}"
+log "DEBUG" "RESTIC_REPOSITORY: ${RESTIC_REPOSITORY}"
+
+log "INFO" "Directory tree:"
+tree -a -P .exclude_from_backup -L 4 ${backup_dir}
+
+log "DEBUG" "Healthcheck start"
 healthcheck /start
 
-# Do not save full backup log to logfile but to backup-last.log
-restic backup ${backup_dir} ${RESTIC_JOB_ARGS} --tag=${RESTIC_TAG?"Missing environment variable RESTIC_TAG"} >> ${lastLogfile} 2>&1
+log "INFO" "Start the restic backup"
+restic backup ${backup_dir} ${RESTIC_JOB_ARGS} --tag=${RESTIC_TAG}
 backupRC=$?
-logLast "Finished backup at $(date)"
+log "INFO" "Finished restic backup"
 if [[ $backupRC == 0 ]]; then
     echo "Backup Successful"
     healthcheck /0
@@ -73,35 +69,39 @@ else
 fi
 
 if [[ $backupRC == 0 ]] && [ -n "${RESTIC_FORGET_ARGS}" ]; then
-    echo "Forget about old snapshots based on RESTIC_FORGET_ARGS = ${RESTIC_FORGET_ARGS}"
-    restic forget ${RESTIC_FORGET_ARGS} >> ${lastLogfile} 2>&1
+    log "INFO" "Forget old snapshots based on RESTIC_FORGET_ARGS = ${RESTIC_FORGET_ARGS}"
+    restic forget ${RESTIC_FORGET_ARGS}
     rc=$?
-    logLast "Finished forget at $(date)"
     if [[ $rc == 0 ]]; then
-        echo "Forget Successful"
+        log "INFO" "Finished restic forget"
     else
-        echo "Forget Failed with Status ${rc}"
+        log "ERROR" "Forget Failed with Status: ${rc}"
         restic unlock
-        copyErrorLog
         healthcheck /fail
     fi
 fi
 
 end=`date +%s`
-echo "Finished Backup at $(date +"%Y-%m-%d %H:%M:%S") after $((end-start)) seconds"
+log "INFO" "Finished Backup at $(date +"%Y-%m-%d %H:%M:%S") after $((end-start)) seconds"
 
 if [ -n "${MAILX_ARGS}" ]; then
-    sh -c "mail -v -S sendwait ${MAILX_ARGS} < ${lastLogfile} > ${lastMailLogfile} 2>&1"
+    log "INFO" "Executing mail command"
+    sh -c "mail -v -S sendwait ${MAILX_ARGS}" 2>&1 | while IFS= read -r line; do
+        log "INFO" "$line"
+    done
     if [ $? == 0 ]; then
-        echo "Mail notification successfully sent."
+        log "INFO" "Mail notification successfully sent."
     else
-        echo "Sending mail notification FAILED. Check ${lastMailLogfile} for further information."
+        log "ERROR" "Sending mail notification FAILED."
     fi
+else
+    log "DEBUG" "MAILX_ARGS not defined. Therefore no mail notification"
 fi
 
+# /hooks/post-backup.sh
 if [ -f "/hooks/post-backup.sh" ]; then
-    echo "Starting post-backup script ..."
+    log "INFO" "Starting post-backup script"
     /hooks/post-backup.sh $backupRC
 else
-    echo "Post-backup script not found ..."
+    log "DEBUG" "Post-backup script not found"
 fi