+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.2.0 (xx.xx.2024)
+
+- LOG_LEVEL and LOG_TYPE added
+- smaller container image
+
 ## v1.1.4 (19.04.2024)
 
 - Cleanup of entry.sh